23/03/26

+++ hardware_processing/queue.h
@@ -20,5 +20,5 @@
 
 PUBLIC void queue_init();
 PUBLIC uint8_t* give_array_address();
-PUBLIC BYTE get_queue_size();
+PUBLIC int get_queue_size();
 PUBLIC bool check_data();
